Remove hardcoded search
